requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
astroid==2.9.3
autopep8==1.6.0
beautifulsoup4==4.10.0
boto3==1.21.27
botocore==1.24.27
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cssselect2==0.5.0
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-heroku==0.3.1
django-storages==1.12.3
django-tempus-dominus==5.1.2.17
Faker==13.3.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
htmldocx==0.0.6
idna==3.3
isort==5.10.1
jmespath==1.0.0
lazy-object-proxy==1.7.1
lxml==4.8.0
mccabe==0.6.1
mysql==0.0.3
mysqlclient==2.1.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
pylint==2.12.2
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
reportlab==3.6.9
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
tqdm==4.63.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.13.3
xhtml2pdf==0.2.6
